LayoutTests/platform/chromium-mac/fast/css/font-face-descriptor-multiple-values-parsing-expected.txt

This tests that the font-weight, font-style and font-variant properties accept value lists when they appear inside @font-face rules, and that the value 'all' is allowed only by itself.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("font-weight", "100") is "100"
PASS test("font-weight", "all") is "all"
PASS test("font-weight", "100, 200") is "100, 200"
PASS test("font-weight", "bold, normal") is "bold, normal"
PASS test("font-weight", "100, 200, 300, 400, 500, 600, 700, 100") is "100, 200, 300, 400, 500, 600, 700, 100"
FAIL test("font-weight", "all, 100") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-weight", "bold, normal, all") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-weight", "") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
PASS test("font-style", "normal") is "normal"
PASS test("font-style", "italic") is "italic"
PASS test("font-style", "normal, oblique") is "normal, oblique"
PASS test("font-style", "all") is "all"
FAIL test("font-style", "all, normal") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-style", "italic, all") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-style", "") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
PASS test("font-variant", "normal") is "normal"
PASS test("font-variant", "small-caps") is "small-caps"
PASS test("font-variant", "normal, small-caps") is "normal, small-caps"
PASS test("font-variant", "all") is "all"
FAIL test("font-variant", "all, normal") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-variant", "small-caps, all") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
FAIL test("font-variant", "") should throw TypeError: 'null' is not an object (evaluating 'style.sheet.rules[0].style.getPropertyCSSValue(property).cssText'). Threw exception TypeError: Cannot read property 'cssText' of null.
PASS successfullyParsed is true

TEST COMPLETE
